pyproject.toml
[project]
name = "pfman-code-exercise"
version = "0.1.0"
description = "AZX, PBC. Portfolio Manager Coding Exercise"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "baml-py>=0.88.0",
    "fastapi[standard]>=0.115.12",
    "h3>=4.2.2",
    "itsdangerous>=2.2.0",
    "loguru>=0.7.3",
    "neomodel[rust-driver-ext]>=5.4.5",
    "prisma>=0.15.0",
    "pycountry>=24.6.1",
    "python-dotenv>=1.1.0",
    "redis>=6.1.0",
    "starlette>=0.46.2",
    "urllib3>=2.4.0",
    "uvicorn[standard]>=0.34.2",
]

[dependency-groups]
dev = ["click>=8.2.0", "pytest>=8.3.5"]


[tool.uv]
package = true

[project.scripts]
graph = "cli.graph_cli:cli"
